refactor: streamline employee get and deletion methods with global exception handling
