+++ employee/src/main/java/io/carrie/employee/common/GlobalExceptionHandler.java
@@ -1,16 +1,17 @@
-package io.carrie.employee.common.exceptions;
+package io.carrie.employee.common;
 
-import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import io.carrie.employee.common.exceptions.NotFoundException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
 
     @ExceptionHandler(NotFoundException.class)
     public ResponseEntity<String> handleNotFound(NotFoundException ex) {
-        return ResponseEntity.notFound().build();
+        return ResponseEntity.status(404).body(ex.getMessage());
     }
 
     @ExceptionHandler(IllegalArgumentException.class)
