+++ employee/src/main/java/io/carrie/employee/common/exceptions/NotFoundException.java
@@ -1,0 +1,11 @@
+package io.carrie.employee.common.exceptions;
+
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.http.HttpStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
